Move reusable components into modules folder
