+++ handlers/discount-api/src/productToDiscountMapping.ts
@@ -1,7 +1,7 @@
 import type { BillingPeriod } from '../../../modules/billingPeriod';
+import { checkDefined } from '../../../modules/nullAndUndefined';
 import type { Stage } from '../../../modules/stage';
-import { checkDefined } from './nullAndUndefined';
-import type { ZuoraSubscription } from './zuora/zuoraSchemas';
+import type { ZuoraSubscription } from '../../../modules/zuora/src/zuoraSchemas';
 
 export const getDiscountFromSubscription = (
 	stage: Stage,
