+++ handlers/discount-api/test/fixtures/request-bodies/digitalSub-subscribe-body-old-price.ts
@@ -1,5 +1,5 @@
 import type { Dayjs } from 'dayjs';
-import { zuoraDateFormat } from '../../../src/zuora/common';
+import { zuoraDateFormat } from '../../../../../modules/zuora/src/common';
 
 export const digiSubSubscribeBody = (
 	subscriptionDate: Dayjs,
