+++ modules/aws/src/config.ts
