+++ modules/catalog/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "catalog",
+  "version": "1.0.0",
+  "scripts": {
+    "test": "jest --group=-integration",
+    "it-test": "jest --group=integration"
+  },
+  "dependencies": {
+    "@aws-sdk/client-s3": "3.451.0",
+    "zod": "^3.22.4"
+  }
+}
