+++ modules/catalog/src/catalogSchema.ts
