+++ modules/catalog/test/catalogIntegration.test.ts
@@ -4,8 +4,8 @@
  * @group integration
  */
 
-import { getCatalogFromS3 } from '../../src/catalog/catalog';
-import type { Catalog } from '../../src/catalog/catalogSchema';
+import { getCatalogFromS3 } from '../src/catalog';
+import type { Catalog } from '../src/catalogSchema';
 
 test('getCatalogFromS3', async () => {
 	const codeCatalog: Catalog = await getCatalogFromS3('CODE');
