+++ modules/zuora/package.json
@@ -1,0 +1,14 @@
+{
+  "name": "zuora",
+  "version": "1.0.0",
+  "scripts": {
+    "test": "jest --group=-integration",
+    "it-test": "jest --group=integration"
+  },
+  "dependencies": {
+    "@aws-sdk/client-s3": "3.451.0",
+    "@aws-sdk/client-secrets-manager": "3.451.0",
+    "dayjs": "^1.11.10",
+    "zod": "^3.22.4"
+  }
+}
