+++ modules/zuora/src/oAuthCredentials.ts
@@ -2,8 +2,8 @@ import {
 	GetSecretValueCommand,
 	SecretsManagerClient,
 } from '@aws-sdk/client-secrets-manager';
-import type { Stage } from '../../../../modules/stage';
-import { awsConfig } from '../aws/config';
+import { awsConfig } from '../../aws/src/config';
+import type { Stage } from '../../stage';
 import type { OAuthClientCredentials } from './zuoraSchemas';
 import { oAuthClientCredentialsSchema } from './zuoraSchemas';
