+++ modules/zuora/src/zuoraClient.ts
@@ -1,5 +1,5 @@
 import type { z } from 'zod';
-import type { Stage } from '../../../../modules/stage';
+import type { Stage } from '../../stage';
 import { BearerTokenProvider } from './bearerTokenProvider';
 import { zuoraServerUrl } from './common';
 import { getOAuthClientCredentials } from './oAuthCredentials';
