+++ package.json
@@ -18,8 +18,9 @@
 		"eslint": "^8.53.0",
 		"hygen": "^6.2.11",
 		"jest": "^29.7.0",
-		"prettier": "^3.1.0",
 		"ts-jest": "^29.1.1",
+		"jest-runner-groups": "^2.2.0",
+		"prettier": "^3.1.0",
 		"typescript": "^5.2.2"
 	}
 }
